+++ ui-ngx/src/app/modules/home/components/table/table-datasource.abstract.ts
@@ -18,7 +18,7 @@ import { DataSource } from '@angular/cdk/collections';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-export abstract class TbDatasource<T> implements DataSource<T> {
+export abstract class TbTableDatasource<T> implements DataSource<T> {
 
   protected dataSubject = new BehaviorSubject<Array<T>>([]);
