--- ui-ngx/src/app/modules/home/components/widget/lib/gateway/connectors-configuration/modbus/public-api.ts
@@ -1,23 +1,0 @@
-///
-/// Copyright © 2016-2024 The Thingsboard Authors
-///
-/// Licensed under the Apache License, Version 2.0 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///     http://www.apache.org/licenses/LICENSE-2.0
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-
-export * from './modbus-basic-config/modbus-basic-config.component';
-export * from './modbus-values/modbus-values.component';
-export * from './modbus-data-keys-panel/modbus-data-keys-panel.component';
-export * from './modbus-slave-config/modbus-slave-config.component';
-export * from './modbus-master-table/modbus-master-table.component';
-export * from './modbus-slave-dialog/modbus-slave-dialog.component';
-export * from './modbus-security-config/modbus-security-config.component';
