--- ui-ngx/src/app/modules/home/components/widget/lib/gateway/connectors-configuration/public-api.ts
@@ -1,27 +1,0 @@
-///
-/// Copyright © 2016-2024 The Thingsboard Authors
-///
-/// Licensed under the Apache License, Version 2.0 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///     http://www.apache.org/licenses/LICENSE-2.0
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-
-export * from './mapping-table/mapping-table.component';
-export * from './device-info-table/device-info-table.component';
-export * from './security-config/security-config.component';
-export * from './opc-server-config/opc-server-config.component';
-export * from './mapping-data-keys-panel/mapping-data-keys-panel.component';
-export * from './type-value-panel/type-value-panel.component';
-export * from './broker-config-control/broker-config-control.component';
-export * from './workers-config-control/workers-config-control.component';
-export * from './opc-ua-basic-config/opc-ua-basic-config.component';
-export * from './mqtt-basic-config/mqtt-basic-config.component';
-export * from './modbus/public-api';
