+++ ui-ngx/src/app/modules/home/pipes/gateway-help-link.pipe.ts
