+++ ui-ngx/src/app/shared/directives/ellipsis-chip-list.directive.ts
